fix manifest generation for pages
