+++ packages/next-swc/crates/next-api/src/lib.rs
@@ -6,6 +6,7 @@ mod dynamic_imports;
 pub mod entrypoints;
 mod font;
 mod instrumentation;
+mod loadable_manifest;
 mod middleware;
 mod pages;
 pub mod paths;
