+++ packages/next-swc/crates/next-api/src/pages.rs
@@ -72,6 +72,7 @@ use crate::{
         DynamicImportedChunks,
     },
     font::create_font_manifest,
+    loadable_manifest::create_react_loadable_manifest,
     paths::{
         all_paths_in_root, all_server_paths, get_js_paths_from_root, get_wasm_paths_from_root,
         wasm_paths_to_bindings,
@@ -749,8 +750,10 @@ impl PageEndpoint {
 
                 let dynamic_import_modules =
                     collect_next_dynamic_imports([Vc::upcast(ssr_module)]).await?;
+                let client_chunking_context =
+                    this.pages_project.project().client_chunking_context();
                 let dynamic_import_entries = collect_evaluated_chunk_group(
-                    edge_chunking_context,
+                    Vc::upcast(client_chunking_context),
                     dynamic_import_modules,
                     Vc::cell(evaluatable_assets.clone()),
                 )
@@ -923,74 +926,20 @@ impl PageEndpoint {
 
     #[turbo_tasks::function]
     async fn react_loadable_manifest(
-        self: Vc<Self>,
+        &self,
         dynamic_import_entries: Vc<DynamicImportedChunks>,
     ) -> Result<Vc<OutputAssets>> {
-        let this = self.await?;
-        let node_root = this.pages_project.project().node_root();
-        let project_src_dir = this.pages_project.pages_dir().parent().await?;
-
-        let dynamic_import_entries = &*dynamic_import_entries.await?;
-
-        let mut output = vec![];
-        let mut loadable_manifest: HashMap<String, LoadableManifest> = Default::default();
-        for (origin, dynamic_imports) in dynamic_import_entries.into_iter() {
-            let origin_path = &*origin.ident().path().await?;
-
-            for (import, chunk_output) in dynamic_imports {
-                let chunk_output = chunk_output.await?;
-                output.extend(chunk_output.iter().copied());
-
-                // https://github.com/vercel/next.js/blob/canary/packages/next/src/build/webpack/plugins/react-loadable-plugin.ts
-                // next_dynamic transform puts relative paths to the project dir
-                // dir for the origin import.
-                let id = format!(
-                    "{} -> {}",
-                    project_src_dir
-                        .get_path_to(origin_path)
-                        .map_or_else(|| origin_path.to_string(), |path| path.to_string()),
-                    import
-                );
-
-                let server_path = node_root.join("server".to_string());
-                let server_path_value = server_path.await?;
-                let files = chunk_output
-                    .iter()
-                    .map(move |file| {
-                        let server_path_value = server_path_value.clone();
-                        async move {
-                            Ok(server_path_value
-                                .get_path_to(&*file.ident().path().await?)
-                                .map(|path| path.to_string()))
-                        }
-                    })
-                    .try_flat_join()
-                    .await?;
-
-                let manifest_item = LoadableManifest {
-                    id: id.clone(),
-                    files,
-                };
-
-                loadable_manifest.insert(id, manifest_item);
-            }
-        }
-
-        let loadable_path_prefix = get_asset_prefix_from_pathname(&this.pathname.await?);
-        let loadable_manifest = Vc::upcast(VirtualOutputAsset::new(
+        let node_root = self.pages_project.project().node_root();
+        let client_relative_path = self.pages_project.project().client_relative_path();
+        let loadable_path_prefix = get_asset_prefix_from_pathname(&self.pathname.await?);
+        Ok(create_react_loadable_manifest(
+            dynamic_import_entries,
+            node_root,
+            client_relative_path,
             node_root.join(format!(
                 "server/pages{loadable_path_prefix}/react-loadable-manifest.json"
             )),
-            AssetContent::file(
-                FileContent::Content(File::from(serde_json::to_string_pretty(
-                    &loadable_manifest,
-                )?))
-                .cell(),
-            ),
-        ));
-
-        output.push(loadable_manifest);
-        Ok(Vc::cell(output))
+        ))
     }
 
     #[turbo_tasks::function]
